Deprecate redundant storage lib. Add promised wrapper for storage get and getSFThread from DB logic. Fixed #48.
